chore: moving helper methods into helpers module
