+++ lib/helpers/helpers.js
@@ -1,0 +1,9 @@
+var noDupes = require('./no-dupes.js')
+var prohibit = require('./prohibit.js')
+var buildName = require('./buildName.js')
+
+module.exports = {
+  prohibit,
+  buildName,
+  noDupes
+}
