+++ lib/helpers/no-dupes.js
@@ -4,6 +4,7 @@
  * @fileoverview Returns a rule that check for duplicated blocks
  * @author Alexander Afanasyev
  */
+
 module.exports = function (kind, branchBlocks, checkedBlocks) {
   function noDupes (context) {
     var suites = []
