+++ lib/helpers/prohibit.js
