+++ lib/rules/missing-expect.js
@@ -5,22 +5,12 @@
  * @author Remco Haszing
  */
 
+var buildName = require('../helpers/helpers').buildName
+
 module.exports = function (context) {
   var allowed = context.options.length ? context.options : ['expect()']
 
   var unchecked = []
-
-  function buildName (node) {
-    if (node.type === 'CallExpression') {
-      return buildName(node.callee) + '()'
-    }
-    if (node.type === 'MemberExpression') {
-      return buildName(node.object) + '.' + node.property.name
-    }
-    if (node.type === 'Identifier') {
-      return node.name
-    }
-  }
 
   return {
     CallExpression: function (node) {
