+++ lib/rules/no-disabled-tests.js
@@ -5,7 +5,7 @@
  * @author Tom Vincent
  */
 
-var prohibit = require('../prohibit')
+var prohibit = require('../helpers/helpers').prohibit
 
 module.exports = function (context) {
   var prohibiteds = [
