+++ lib/rules/no-expect-in-setup-teardown.js
@@ -5,24 +5,14 @@
  * @author Alexander Afanasyev
  */
 
+var buildName = require('../helpers/helpers').buildName
+
 module.exports = function (context) {
   var allowed = context.options.length ? context.options : ['expect()']
   var setupRegexp = /^(before|after)(Each|All)$/
 
   var insideSetup = false
   var setupFunctionName
-
-  function buildName (node) {
-    if (node.type === 'CallExpression') {
-      return buildName(node.callee) + '()'
-    }
-    if (node.type === 'MemberExpression') {
-      return buildName(node.object) + '.' + node.property.name
-    }
-    if (node.type === 'Identifier') {
-      return node.name
-    }
-  }
 
   return {
     CallExpression: function (node) {
