+++ lib/rules/no-spec-dupes.js
@@ -13,4 +13,4 @@ var checkedBlocks = [
   'it'
 ]
 
-module.exports = require('../no-dupes')('spec', branchBlocks, checkedBlocks)
+module.exports = require('../helpers/helpers').noDupes('spec', branchBlocks, checkedBlocks)
