+++ lib/rules/no-suite-dupes.js
@@ -13,4 +13,4 @@ var checkedBlocks = [
   'describe'
 ]
 
-module.exports = require('../no-dupes')('suite', branchBlocks, checkedBlocks)
+module.exports = require('../helpers/helpers').noDupes('suite', branchBlocks, checkedBlocks)
